[REFACTORING] Split IMAPServerCompressTest
